null check
git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk@3515 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab

+++ nexus-indexer/src/main/java/org/sonatype/nexus/index/context/DefaultIndexingContext.java
@@ -267,9 +267,12 @@ public class DefaultIndexingContext
     {
         String[] names = indexDirectory.list();
 
-        for ( int i = 0; i < names.length; i++ )
+        if ( names != null )
         {
-            indexDirectory.deleteFile( names[i] );
+            for ( int i = 0; i < names.length; i++ )
+            {
+                indexDirectory.deleteFile( names[i] );
+            }
         }
 
         IndexUtils.deleteTimestamp( indexDirectory );
